+++ frontend/src/InteractiveImport/Interactive/InteractiveImportModalContent.js
@@ -294,7 +294,6 @@ class InteractiveImportModalContent extends Component {
       isFetching,
       isPopulated,
       isSaving,
-      isCombineInProgress,
       error,
       items,
       sortKey,
@@ -529,7 +528,7 @@ class InteractiveImportModalContent extends Component {
 
             <Button
               kind={kinds.SUCCESS}
-              isDisabled={isSaving || isCombineInProgress || !selectedIds.length || !!invalidRowsSelected.length || inconsistentBookReleases}
+              isDisabled={isSaving || !selectedIds.length || !!invalidRowsSelected.length || inconsistentBookReleases}
               onPress={this.onImportSelectedPress}
             >
               Import
@@ -620,7 +619,6 @@ InteractiveImportModalContent.propTypes = {
   sortKey: PropTypes.string,
   sortDirection: PropTypes.string,
   interactiveImportErrorMessage: PropTypes.string,
-  isCombineInProgress: PropTypes.bool,
   onSortPress: PropTypes.func.isRequired,
   onFilterExistingFilesChange: PropTypes.func.isRequired,
   onReplaceExistingFilesChange: PropTypes.func.isRequired,